+++ src/core/shared/enums/system.ts
@@ -53,6 +53,8 @@ export enum SYSTEM_EVENTS {
     PLAYER_EMIT_NOTIFICATION = 'notification:Show',
     PLAYER_EMIT_SPINNER = 'spinner:Show',
     PLAYER_EMIT_SPINNER_CLEAR = 'spinner:Clear',
+    PLAYER_EMIT_SHARD = 'shard:Create',
+    PLAYER_EMIT_SHARD_CLEAR = 'shard:Clear',
     PLAYER_EMIT_TASK_MOVE = 'task:Move',
     PLAYER_EMIT_TASK_TIMELINE = 'task:Timeline',
     //